cleanup standard text fixture imports
