+++ news/521.api_change
@@ -1,0 +1,1 @@
+Plugins should now include test fixtures (sweep_runner, task_runner) via a standardized conftest.py
