+++ plugins/examples/example_launcher_plugin/tests/test_example_launcher_plugin.py
@@ -7,10 +7,6 @@ from hydra.test_utils.launcher_common_tests import (
     IntegrationTestSuite,
     LauncherTestSuite,
 )
-
-# This has to be included here for the LauncherTestSuite to work.
-# noinspection PyUnresolvedReferences
-from hydra.test_utils.test_utils import sweep_runner  # noqa: F401
 from hydra_plugins.example_launcher_plugin.example_launcher import ExampleLauncher
