+++ plugins/examples/example_searchpath_plugin/tests/conftest.py
@@ -1,2 +1,3 @@
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
+# Add Hydra standard test fixtures
 from hydra.test_utils._conftest import *  # noqa F401
