+++ plugins/examples/example_searchpath_plugin/tests/test_example_search_path_plugin.py
@@ -2,12 +2,7 @@
 from hydra.core.global_hydra import GlobalHydra
 from hydra.core.plugins import Plugins
 from hydra.plugins.search_path_plugin import SearchPathPlugin
-
-# noinspection PyUnresolvedReferences
-from hydra.test_utils.test_utils import (  # noqa: F401
-    TGlobalHydraContext,
-    hydra_global_context,
-)
+from hydra.test_utils.test_utils import TGlobalHydraContext
 from hydra_plugins.example_searchpath_plugin.example_searchpath_plugin import (
     ExampleSearchPathPlugin,
 )
@@ -20,9 +15,7 @@ def test_discovery() -> None:
     ]
 
 
-def test_config_installed(
-    hydra_global_context: TGlobalHydraContext,  # noqa: F811
-) -> None:
+def test_config_installed(hydra_global_context: TGlobalHydraContext) -> None:
     with hydra_global_context():
         config_loader = GlobalHydra.instance().config_loader()
         assert "my_default_output_dir" in config_loader.get_group_options(
